Fix automatic launch of the APP container.
Currently when running the containers with:

    $ docker-compose -f dci.yml -f dci-db_init.yml up

We ends up with the following message error:

app_1        | [23:11:28] Local gulp not found in /opt/dci-ui
app_1        | [23:11:28] Try running: npm install gulp
dcidevenv_app_1 exited with code 1

The node_modules folder is created inside the image by the Dockerfile but it
does not exist in the Docker host. The result is that the volume binding
"./dci-ui:/opt/dci-ui" will hides the node_modules content within the
container. In fact the node_modules folder is empty which cause gulp command
to not be found.

Fortunately, the content of node_modules is not lost, indeed Docker
created a volume copy of /opt/dci-ui/nodes_modules so we just need to
mount it in the container to make it visible again at running time. This is
the purpose of adding the volume /opt/dci-ui/node_modules.

Change-Id: I2b2cb56cda33d134f446dfefd59253d204dc2eb3