dbwatcher: fix postgresql_autodoc command
There is no package on centos 7 for postgresql_autodoc. The current
workaround is to use the rpm from fedora instead.
Then the pgpass is not usefull there because this version doesn't
support pgpass file.
Finally we never get the database ip address to use in the autodoc
command.
To get this working we also need a dependency on the API because
this one runs the db provisionning script.

Change-Id: I4d33b42c2066c248496ae3b4d2c8af9588f7c314