TOX: Use the current code and not master from git
Currently when we run tox in /opt/dci-client it goes and install
dci-control-server from upstream/master rather than the one with the
changes applied on the dci-dev-env environment.

This commit aims to force tox to use the current installation of
dci-control-server rather than the one from upstream/master. This way we
can reproduce the exact same result locally that the one in the CI.

Change-Id: I694427b21ded58a69e1bc5dcef5bda4298b2e4a4