app: install deps at run time
Until now, the npm modules were deployed during the image constructuion.
This solution comes with a bunch a of drawback:
- we need to rebuild the image as soon as the dep are changed
- we have to manually trigger ther rebuild
- we have to manage the node_modules directory differently

This commit changes a bit the strategy. It relies on the `yarn` command
to pull the dependency because it's way fast than `npm`. This allow us
to check the stats of the dependency before we start the container. If
we introduce a new dep, we just have to restart the container.

Change-Id: Iaf15b50c985736e6fb35f3d6a50eabf287c93f40