Load dumps into dci-dev-env
Purpose of this commit is to easily be able to load database dumps into
the dci-dev-env environment.

This way developpers can easily load the production dump directly on
their dci-dev-env setup and work with real data. It also help when there
is a need to solve a bug only present in production.

Change-Id: Idadc598cef1ad0f80419847d74da8997d6675eb8