Use pg_restore from the db container
Since we already have all the tools in the database container it's
better to use them instead of relying on the host to have them.

Download the backup to the `dci-db` folder to make it accessible from
inside the container.

Show some stats about the restored backup after pg_restore is done.

Change-Id: I0c542801eb3ceed6e218fad4964c2ea40647dd56